remove speaker context
