+++ src/contexts/Provider.jsx
@@ -16,7 +16,6 @@ import { AlertsProvider } from "./AlertsContext.jsx";
 import { EventProvider } from "./EventContext.jsx";
 import { MediaProvider } from "./MediaContext.jsx";
 import { SponsorsProvider } from "./SponsorsContext.jsx";
-import SpeakerProvider from "./SpeakerContext.jsx";
 
 export const Provider = ({ children }) => {
   return (
@@ -40,13 +39,11 @@ export const Provider = ({ children }) => {
                         <PagesProvider>
                           <AlertsProvider>
                             <EventProvider>
-                              <SpeakerProvider>
-                                <MediaProvider>
-                                  <Router>
-                                    {children}
-                                  </Router>
-                                </MediaProvider>
-                              </SpeakerProvider>
+                              <MediaProvider>
+                                <Router>
+                                  {children}
+                                </Router>
+                              </MediaProvider>
                             </EventProvider>
                           </AlertsProvider>
                         </PagesProvider>
